requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
jinja2==3.1.3
Pillow==9.0.1
matplotlib
boto3

